chore: rename file to personal info
